Fix all CI compiler warnings (#667)
* Fix clang 12 warnings

* Fix windows x86_64 warnings

* Fix linux x86 warnings

+++ src/game/g_main.cpp
@@ -590,7 +590,7 @@ This must be the very first function compiled into the .q3vm file
 ================
 */
 
-extern "C" FN_PUBLIC int vmMain(int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6)
+extern "C" FN_PUBLIC intptr_t vmMain(int command, intptr_t arg0, intptr_t arg1, intptr_t arg2, intptr_t arg3, intptr_t arg4, intptr_t arg5, intptr_t arg6)
 {
 	switch (command)
 	{
